+++ src/kafka/helpers/unmarshallers.test.ts
@@ -18,6 +18,7 @@ const mockPrivateFor = ['testPrivateFor']
 const mockPrivacyGroupId = 'testPrivateGroupID'
 export const mockTxResponse = {
   id: 'id',
+  jobId: 'jobId',
   transaction: {
     from: mockFrom,
     contract: 'myContract[contractTag]',
@@ -62,6 +63,7 @@ describe('unmarshallers', () => {
     it('should unmarshal an envelope successfully', () => {
       const expectedValue: IResponseValue = {
         id: mockTxResponse.id,
+        jobId: mockTxResponse.jobId,
         errors: mockTxResponse.errors,
         contextLabels: mockTxResponse.contextLabels,
         receipt: {